refactor: Api모듈 인수 테스트 통합
